==> Scripts/02_TOCcorrection.R <==
#### TOC transformations ####

TOC_trial1 <- c(0.231756032,
                 0.161029217,
                 0.10172882,
                 0.184820568,
                 0.759227707,
                 0.210970479,
                 0.338475613,
                 1.282422112,
                 0.367973314,
                 0.196905938,
                 0.205112737,
                 0.181027549,
                 0.192837579,
                 2.244962061,
                 1.219245653,
                 0.912238431,
                 0.460681324,
                 0.093583616)

TOC_trial2 <- c(6.306294875,
6.512819153,
6.092996221,
6.704744766,
6.011051649,
6.60321384,
6.159038072,
7.738650125,
5.204791517,
5.99235913,
7.302733281,
6.679713455,
4.881368811,
5.889701751,
5.353836422,
4.815991522,
6.418590548,
6.538591186)

# plot(first_trial,second_trial)
# 
# equation <- lm(second_trial~first_trial)
# abline(equation, col = "red")
# summary(equation)

diffTOC <- c(TOC_trial2-TOC_trial1)
plot(diffTOC)
mean(diffTOC)
sd(diffTOC)

MBC_trial1 <- c(0.154225282,
                0.041735432,
                0.004981098,
                0.028876212,
                1.341792479,
                0.127131405,
                0.339740067,
                2.509454953,
                0.073709172,
                0.206726971,
                0.350122915,
                0.266020136,
                1.02613603,
                3.823029014,
                2.56943762,
                0.897895961,
                0.894133477,
                0.112170459)

MBC_trial2 <- c(1.75828576,
                5.270994458,
                0.155199743,
                2.170516266,
                1.567276765,
                0.998976888,
                1.213703157,
                3.486506569,
                1.803897823,
                1.378234155,
                2.37653623,
                2.415008805,
                3.702472584,
                4.787654619,
                2.757157332,
                0.798911974,
                4.273774885,
                4.587479883)

diffMBC <- c(MBC_trial2-MBC_trial1)
plot(diffMBC)
mean(diffMBC)
sd(diffMBC)
